+++ Editor/ItemCreationWizardEditorWindow.cs
@@ -6,8 +6,6 @@ using UnityEngine.UIElements;
 public class ItemCreationWizardEditorWindow : EditorWindow
 {
     [SerializeField] private VisualTreeAsset _uxml = default;
-    private ItemCreationWizardEditor _itemCreationWizard;
-
 
     [MenuItem("Window/Slax/Inventory/Item Creation Wizard")]
     public static void OpenWindow()
@@ -20,7 +18,6 @@ public class ItemCreationWizardEditorWindow : EditorWindow
     {
         VisualElement tree = _uxml.Instantiate();
         rootVisualElement.Add(tree);
-        _itemCreationWizard = new ItemCreationWizardEditor();
-        _itemCreationWizard.Setup(rootVisualElement);
+        new ItemCreationWizardEditor(rootVisualElement);
     }
 }